# docs/swagger.yaml
basePath: /
definitions:
...
  httpapi.connectResponse:
    properties:
      connection:
        $ref: '#/definitions/httpapi.pairingConnection'
    type: object
...
  httpapi.envelope:
    properties:
      data: {}
    type: object
...
  httpapi.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/httpapi.errorBody'
    type: object
...
  httpapi.instanceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/httpapi.instanceEnvelope'
        type: array
    type: object
...
  httpapi.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  httpapi.meEnvelope:
    properties:
      me:
        $ref: '#/definitions/httpapi.identityResponse'
    type: object
...
  inbound.Payload:
    properties:
      conversation:
        $ref: '#/definitions/inbound.Conversation'
      event:
        type: string
      message:
        $ref: '#/definitions/inbound.Message'
    type: object
...
info:
  contact: {}
  description: 'Standalone multi-instance WhatsApp gateway with authenticated REST
    commands and durable events. Normal REST JSON successes use {"data": ...}; standard
    errors use {"error": {"code", "message"}}. No-content responses, binary media,
    the raw Chatwoot webhook acknowledgement and Manager HTML/redirects have their
    own formats. Dual-auth API routes accept a valid wzap_session cookie or the apikey
    header. Global keys and admin sessions have administrative scope; user sessions
    are restricted to owned instances; instance keys reach only their own instance
    and cannot use collection/admin routes. Swagger 2.0 cannot declare cookie security:
    use an existing manager session cookie for session calls, or Authorize with an
    apikey for machine calls. The /auth/* endpoints use session cookies, never apikey
    authentication; login, logout, health probes, Swagger, Manager and the Chatwoot
    webhook load without an apikey.'
  title: wzap
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies the email/password credentials and mints the wzap_session
        cookie. Unknown email and wrong password share one 401 response.
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Identity, wrapped in the data envelope; the wzap_session cookie
            is set
          headers:
            Set-Cookie:
              description: Sets the httpOnly wzap_session cookie
              type: string
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.envelope'
            - properties:
                data:
                  $ref: '#/definitions/httpapi.meEnvelope'
              type: object
        "400":
          description: Malformed body
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "401":
          description: Invalid credentials
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "413":
          description: Body exceeds the 1 MiB limit
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "429":
          description: Login rate limit exceeded
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      summary: Log in to the manager session
      tags:
      - auth
...
  /auth/me:
    get:
      description: Answers the identity of the wzap_session cookie holder, 401 without
        a valid session.
      produces:
      - application/json
      responses:
        "200":
          description: Identity, wrapped in the data envelope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.envelope'
            - properties:
                data:
                  $ref: '#/definitions/httpapi.meEnvelope'
              type: object
        "401":
          description: Missing or invalid session
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      summary: Show the current manager session
      tags:
      - auth
  /chatwoot/webhook/{id}:
    post:
      consumes:
      - application/json
      description: 'Public by design: no apikey or session cookie is required. The
        per-instance limiter can return 429. Success and discarded events return the
        raw {"content":""} acknowledgement, outside the REST data envelope. Operational
        commands are discarded here and must use the authenticated /instances/{id}/chatwoot/command
        route.'
      parameters:
      - description: Instance UUID or name (exact, case-sensitive)
        in: path
        name: id
        required: true
        type: string
      - description: Chatwoot event subset; unknown fields are ignored
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.Payload'
      produces:
      - application/json
      responses:
        "200":
          description: Raw acknowledgement with empty content
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            properties:
              content:
                type: string
            type: object
        "400":
          description: Connector disabled or invalid body (including oversized body)
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "404":
          description: Invalid or missing instance
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "409":
          description: 'instance_name_ambiguous: legacy name matches multiple instances'
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "429":
          description: Per-instance webhook rate limit exceeded
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      summary: Receive a Chatwoot webhook
      tags:
      - chatwoot
...
  /instances:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: All authorized instances, wrapped in the data envelope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.envelope'
            - properties:
                data:
                  $ref: '#/definitions/httpapi.instanceListResponse'
              type: object
        "401":
          description: Missing or invalid credential
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "403":
          description: Instance keys own no collection view
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      security:
      - apikey: []
      summary: List instances
      tags:
      - instances
    post:
      consumes:
      - application/json
      description: 'Answers with the aggregated instance DTO: the webhook block nests
        under integration and the nullable settings blocks under settings (both null
        on a fresh instance except the persisted default_disappearing echo).'
      parameters:
      - description: Instance payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.createInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created instance plus its one-time key, wrapped in the data
            envelope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.envelope'
            - properties:
                data:
                  $ref: '#/definitions/httpapi.createInstanceResponse'
              type: object
        "400":
          description: Malformed body
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "401":
          description: Missing or invalid credential
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "403":
          description: Forbidden or quota exceeded
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "409":
          description: 'instance_name_taken: name already taken, or external ref already
            taken'
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "413":
          description: Body exceeds the 1 MiB limit
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "422":
          description: 'invalid_instance_name: invalid/reserved name; unknown owner
            or invalid webhook config'
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      security:
      - apikey: []
      summary: Create an instance
      tags:
      - instances
...
  /instances/{id}/apikey:
    delete:
      parameters:
      - description: Instance UUID or name (exact, case-sensitive)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Revoked, no body
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
        "401":
          description: Missing or invalid credential
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "403":
          description: Requires global or admin scope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "404":
          description: Instance not found
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "409":
          description: 'instance_name_ambiguous: legacy name matches multiple instances'
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      security:
      - apikey: []
      summary: Revoke an instance API key
      tags:
      - apikeys
  /instances/{id}/apikey/rotate:
    post:
      parameters:
      - description: Instance UUID or name (exact, case-sensitive)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Fresh key, wrapped in the data envelope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.envelope'
            - properties:
                data:
                  $ref: '#/definitions/httpapi.rotateAPIKeyResponse'
              type: object
        "401":
          description: Missing or invalid credential
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "403":
          description: Requires global or admin scope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "404":
          description: Instance not found
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "409":
          description: 'instance_name_ambiguous: legacy name matches multiple instances'
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      security:
      - apikey: []
      summary: Rotate an instance API key
      tags:
      - apikeys
...
  /instances/{id}/connect:
    post:
      parameters:
      - description: Instance UUID or name (exact, case-sensitive)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pairing result, wrapped in the data envelope
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.envelope'
            - properties:
                data:
                  $ref: '#/definitions/httpapi.connectResponse'
              type: object
        "401":
          description: Missing or invalid credential
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "403":
          description: Not the owner
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "404":
          description: Instance not found
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "409":
          description: Instance already connected; instance_name_ambiguous for a legacy
            name matching multiple instances
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
        "500":
          description: Internal error
          headers:
            X-Request-Id:
              description: Correlation id, generated when absent
              type: string
          schema:
            $ref: '#/definitions/httpapi.errorEnvelope'
      security:
      - apikey: []
      summary: Connect an instance
      tags:
      - connection
...
securityDefinitions:
  apikey:
    description: 'Machine credential in the literal apikey header: a global key grants
      administrative scope, an instance key grants access only to its own instance.
      Dual-auth API routes also accept the wzap_session cookie (user ownership/admin
      scope); cookie security cannot be modeled in Swagger 2.0. /auth/* uses cookies
      rather than this header.'
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"